requirements.txt
torch>=2.0.0
transformers>=4.20.0
numpy<2
pandas
scipy>=1.7.0
tqdm
pyyaml